+++ tosho_kmkc/tests/descramble_test.rs
@@ -1,18 +1,47 @@
 use std::{
     fs::File,
     io::{Read, Seek},
+    path::PathBuf,
+    str::FromStr,
 };
 
 use tosho_kmkc::imaging::descramble_image;
+
+fn open_assets_file(file_name: &str) -> Option<File> {
+    let manifest_dir = PathBuf::from_str(env!("CARGO_MANIFEST_DIR")).unwrap();
+    let root_dir = manifest_dir.parent().unwrap();
+
+    let assets_dir = root_dir.join("tosho_assets");
+
+    if !assets_dir.exists() {
+        return None;
+    }
+
+    if !assets_dir.is_dir() {
+        return None;
+    }
+
+    let file_path = assets_dir.join("kmkc").join(file_name);
+
+    if !file_path.exists() {
+        return None;
+    }
+
+    File::open(file_path).ok()
+}
 
 #[test]
 fn test_descramble_image() {
     let seed = 749191485_u32;
     let rectbox = 4_u32;
 
-    let manifest_dir = std::env::var("CARGO_MANIFEST_DIR").unwrap();
-    let mut img_file =
-        File::open(format!("{}/tests/descramble_src.tmfxture", manifest_dir)).unwrap();
+    let mut img_file = match open_assets_file("descramble_src.tmfxture") {
+        Some(file) => file,
+        None => {
+            assert!(true, "File not found, skipping test");
+            return;
+        }
+    };
 
     // Check if file exists, if not skip the test (lfs not fetched or something)
     if img_file.metadata().is_err() {
@@ -50,8 +79,7 @@ fn test_descramble_image() {
     let descrambled = descramble_image(buf.as_ref(), rectbox, seed).unwrap();
 
     // Test with the reference image
-    let mut ref_file =
-        File::open(format!("{}/tests/descramble_out.tmfxture", manifest_dir)).unwrap();
+    let mut ref_file = open_assets_file("descramble_out.tmfxture").unwrap();
     let mut ref_buf = vec![];
     ref_file
         .read_to_end(&mut ref_buf)